Correct statement about role of OAuth
While this is an "advanced" concept to those new to OAuth, it's important the Mulesoft, not mis-educate users about the role of OAuth.

The example, while simple, leads users to believe OAuth can or should be used not only for delegated authorization but also for authentication.
The example first suggests this with "Sign in with Tweetbooks" and then also suggests the access token represents an authentication.
In fact, this is only correct in a happy path, but not when you have a malicious attacker. 
This is the very problem OpenID Connect is solving - as a way of discouraging the bad security practice of applications using IdPs like Facebook and Twitter for the purpose of authenticating to their own apps. OpenID connect issue an ID Token (JWT) to provide this to the Relying Party.

Btw ... the OAuth provider example needs to be corrected to reflect the HTTP connector for Mule 3.7.x
And also, the OAuth consumer does not run from within the IDE (the JSPs don't compile complaining that a class is a directory - which typically means it cannot find a class on the classpath).